+++ src/test/java/com/global/api/tests/gpapi/GpApiReportingDisputesTest.java
@@ -16,6 +16,7 @@ import com.global.api.entities.reporting.SearchCriteria;
 import com.global.api.serviceConfigs.GpApiConfig;
 import com.global.api.services.ReportingService;
 import com.global.api.utils.DateUtils;
+import org.joda.time.DateTime;
 import org.junit.FixMethodOrder;
 import org.junit.Ignore;
 import org.junit.Test;
@@ -1006,4 +1007,33 @@ public class GpApiReportingDisputesTest extends BaseGpApiReportingTest {
         }
     }
 
+    @Test
+    public void ReportDisputeDetail_stageTime() throws ApiException {
+        String disputeId = "DIS_SAND_abcd1234";
+        String dateFormat = "dd-MM-yyyy";
+        String expectedStageDate = DateUtils.toString(DateTime.now().minusDays(7).toDate(), dateFormat);
+
+        DisputeSummary response =
+                ReportingService
+                        .disputeDetail(disputeId)
+                        .execute();
+
+        assertNotNull(response);
+        assertEquals(disputeId, response.getCaseId());
+        assertEquals(expectedStageDate, DateUtils.toString(response.getCaseStageTime().toDate(), dateFormat));
+    }
+
+    @Test
+    public void ReportDisputeDetail_brandReference() throws ApiException {
+        String disputeId = "DIS_SAND_abcd1234";
+
+        DisputeSummary response =
+                ReportingService
+                        .disputeDetail(disputeId)
+                        .execute();
+
+        assertNotNull(response);
+        assertEquals(disputeId, response.getCaseId());
+        assertEquals("23423421342323A", response.getTransactionBrandReference());
+    }
 }